# api-gateway-exam/api/handlers/models/comment.go
package models

type Comment struct {
	ID        string `json:"id"`
	PostID    string `json:"post_id"`
	UserID    string `json:"user_id"`
	Content   string `json:"content"`
	CreatedAt string `json:"created_at"`
	UpdatedAt string `json:"updated_at"`
}

type ListComments struct {
	Count    int64      `json:"count"`
	Comments []*Comment `json:"comments"`
}